Send TERM to exec processes before sending KILL signal
